Fix 107322: MacOS/ATI Crash When Starting Eevee
When using ATI GPU on MacOS/Metal compilation of Eevee shaders crashed due to an infinitive
loop in the compiler backend. This change fixes this by using a more reliable loop that
ensures that the loop will be finite.

Thanks for the Apple developers working on finding the root cause and Michael Parking White
for the initial fix.

Fixes #107322
Fixes #107451

+++ source/blender/draw/engines/eevee/shaders/ambient_occlusion_lib.glsl
@@ -118,7 +118,7 @@ float search_horizon(vec3 vI,
   }
 
   float prev_time, time = 0.0;
-  for (float iter = 0.0; time < ssray.max_time && iter < sample_count; iter++) {
+  for (int iter = 0; time < ssray.max_time && iter < sample_count; iter++) {
     prev_time = time;
     /* Gives us good precision at center and ensure we cross at least one pixel per iteration. */
     time = 1.0 + iter + sqr((iter + noise) / sample_count) * ssray.max_time;
