+++ source/blender/draw/engines/eevee/shaders/raytrace_lib.glsl
@@ -137,8 +137,8 @@ bool raytrace(Ray ray,
 #ifdef METAL_AMD_RAYTRACE_WORKAROUND
   bool hit_failsafe = true;
 #endif
-  const float max_steps = 255.0;
-  for (float iter = 1.0; !hit && (time < ssray.max_time) && (iter < max_steps); iter++) {
+  const int max_steps = 255;
+  for (int iter = 1; !hit && (time < ssray.max_time) && (iter < max_steps); iter++) {
     float stride = 1.0 + iter * params.trace_quality;
     float lod = log2(stride) * lod_fac;
 
@@ -210,8 +210,8 @@ bool raytrace_planar(Ray ray, RayTraceParameters params, int planar_ref_id, out 
   /* On very sharp reflections, the ray can be perfectly aligned with the view direction
    * making the tracing useless. Bypass tracing in this case. */
   bool hit = false;
-  const float max_steps = 255.0;
-  for (float iter = 1.0; !hit && (time < ssray.max_time) && (iter < max_steps); iter++) {
+  const int max_steps = 255;
+  for (int iter = 1; !hit && (time < ssray.max_time) && (iter < max_steps); iter++) {
     float stride = 1.0 + iter * params.trace_quality;
 
     prev_time = time;
